override inquirer module methods
