--- forked_inquirer/inquirer/render/console/_path.py
@@ -1,7 +1,0 @@
-# -*- coding: utf-8 -*-
-
-from . import Text
-
-
-class Path(Text):
-    pass
